api/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-huggingface>=0.0.1
langchain-community>=0.0.20
faiss-cpu>=1.7.4
sentence-transformers>=2.2.0
